+++ app/src/main/java/com/fibelatti/photowidget/platform/AppTheme.kt
@@ -21,6 +21,7 @@ fun AppTheme(
 
     ExtendedTheme(
         dynamicColor = userPreferences.dynamicColors,
+        useTrueBlack = userPreferences.useTrueBlack,
         content = content,
     )
 }