+++ app/src/main/java/com/fibelatti/photowidget/preferences/UserPreferencesStorage.kt
@@ -30,6 +30,7 @@ class UserPreferencesStorage @Inject constructor(@ApplicationContext context: Co
     private val _userPreferences = MutableStateFlow(
         UserPreferences(
             appearance = appearance,
+            useTrueBlack = useTrueBlack,
             dynamicColors = dynamicColors,
             defaultSource = defaultSource,
             defaultShuffle = defaultShuffle,
@@ -51,6 +52,15 @@ class UserPreferencesStorage @Inject constructor(@ApplicationContext context: Co
         set(value) {
             sharedPreferences.edit { putString(Preference.APP_APPEARANCE.value, value.name) }
             _userPreferences.update { current -> current.copy(appearance = value) }
+        }
+
+    var useTrueBlack: Boolean
+        get() {
+            return sharedPreferences.getBoolean(Preference.USE_TRUE_BLACK.value, false)
+        }
+        set(value) {
+            sharedPreferences.edit { putBoolean(Preference.USE_TRUE_BLACK.value, value) }
+            _userPreferences.update { current -> current.copy(useTrueBlack = value) }
         }
 
     var dynamicColors: Boolean
@@ -236,6 +246,7 @@ class UserPreferencesStorage @Inject constructor(@ApplicationContext context: Co
         _userPreferences.update {
             UserPreferences(
                 appearance = appearance,
+                useTrueBlack = useTrueBlack,
                 dynamicColors = dynamicColors,
                 defaultSource = defaultSource,
                 defaultShuffle = defaultShuffle,
@@ -250,6 +261,7 @@ class UserPreferencesStorage @Inject constructor(@ApplicationContext context: Co
 
     private enum class Preference(val value: String) {
         APP_APPEARANCE(value = "user_preferences_appearance"),
+        USE_TRUE_BLACK(value = "user_preferences_use_true_black"),
         APP_DYNAMIC_COLORS(value = "user_preferences_dynamic_colors"),
 
         DEFAULT_ASPECT_RATIO(value = "default_aspect_ratio"),