+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * New tap action to pause/resume automatic photo cycling
+* New app theme option for using a true black background
 
 ### Changed